Added Lombok and reworked AbstractModule.java
